+++ server/main.go
@@ -4,19 +4,44 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
 )
 
-// func getUniversities(c *gin.Context) {
-// 	db, _ := gorm.Open(sqlite.Open("unigenie.db"), &gorm.Config{})
-// 	// var universities []map[string]interface{}
-// 	// db.Table("University", " ").Find(&universities)
+type University struct {
+	gorm.Model
+	ID      uint   `gorm:"primaryKey;autoIncrement" json:"university ID"`
+	Name    string `json:"name"`
+	Ranking uint   `json:"ranking"`
+	Country string `json:"country"`
+}
 
-// 	universities := db.Table("University").Select("name")
-// 	fmt.Println(universities)
+type User struct {
+	gorm.Model
+	ID       uint   `gorm:"primaryKey;autoIncrement" json:"User ID"`
+	Name     string `json:"name"`
+	Username string `json:"username"`
+	Password string `json:"password"`
+}
 
-// 	// db.Find(&University)
-// 	c.IndentedJSON(http.StatusOK, universities)
-// }
+type OfferedCourse struct {
+	gorm.Model
+	ID         uint   `gorm:"primaryKey;autoIncrement" json:"Course ID"`
+	CourseName string `json:"course_name"`
+}
+
+func getUniversities(c *gin.Context) {
+	db, err := gorm.Open(sqlite.Open("unigenie.db"), &gorm.Config{})
+	if err != nil {
+		panic("failed to connect database")
+	}
+
+	universities := []University{}
+
+	db.Find(&universities)
+	c.JSON(200, &universities)
+
+}
 
 func setupRouter() *gin.Engine {
 	// Disable Console Color
@@ -32,36 +57,9 @@ func setupRouter() *gin.Engine {
 
 func main() {
 	r := setupRouter()
-	setDatabase()
 
-	// type University struct {
-	// 	gorm.Model
-	// 	UniversityID uint   `gorm:"primaryKey;autoIncrement" json:"university ID"`
-	// 	Name         string `json:"name"`
-	// 	Ranking      uint   `json:"ranking"`
-	// 	Country      string `json:"country"`
-	// 	// CreatedAt    time.Time `json:"createdAt"`
-	// 	// UpdatedAt    time.Time `json:"updatedAt"`
-	// }
+	r.GET("/universities", getUniversities)
 
-	// db, _ := gorm.Open(sqlite.Open("unigenie.db"), &gorm.Config{})
-	// var universities []map[string]interface{}
-	// db.Table("University", " ").Find(&universities)
-
-	// var univ University
-	// res := map[string]interface{}{}
-
-	// fmt.Println("")
-	// fmt.Println("###########################                   #####################################")
-	// rows, _ := db.Table("Universities").Select("name").Rows()
-	// db.Table("Universities").Take(&res)
-	// for rows.Next() {
-	// 	x := rows.Scan()
-	// 	fmt.Println(x)
-	// }
-	// fmt.Println("###########################                   #####################################")
-	// fmt.Println("")
-	// r.GET("/universities", getUniversities)
 	// Listen and Server in 0.0.0.0:8080
 	r.Run(":8080")
 }
